+++ create.sh
@@ -6,6 +6,6 @@ docker rmi -f petrjahoda/linaplast_userimport:latest
 docker build -t petrjahoda/linaplast_userimport:latest .
 docker push petrjahoda/linaplast_userimport:latest
 
-docker rmi -f petrjahoda/linaplast_userimport:2020.3.3
-docker build -t petrjahoda/linaplast_userimport:2020.3.3 .
-docker push petrjahoda/linaplast_userimport:2020.3.3
+docker rmi -f petrjahoda/linaplast_userimport:2021.1.2
+docker build -t petrjahoda/linaplast_userimport:2021.1.2 .
+docker push petrjahoda/linaplast_userimport:2021.1.2