+++ CHANGELOG.md
@@ -14,6 +14,12 @@ adhere to [Semantic Versioning](http://semver.org/spec/v2.0.0.html) rules.
 * ```Fixed``` for any bug fixes.
 * ```Security``` in case of vulnerabilities.
 
+
+## [2021.2.1.1] - 2021-2-1
+
+### Fixed
+- when updating user, setting - when no EVOLOGIN FOUND
+
 ## [2020.3.3.9] - 2020-9-9
 
 ### Fixed